api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.7
python-jose==3.3.0
